Chat server application using java sockets (with GUI)
Designed and developed a multi-user chat application with chat/shared
whiteboard system in java using socket programming.
The application allows users to chat and draw on the whiteboard to
interact with online members in real time. It transmits the messages
through socket and also archives the client’s chat history along with
their identity.
The application was successfully run on AFS server and can interact with
the clients running on windows and mac.